rename proxy package to seekable
