+++ repository_test.go
@@ -6,7 +6,7 @@ import (
 
 	"gopkg.in/src-d/go-git.v3/clients/http"
 	"gopkg.in/src-d/go-git.v3/core"
-	"gopkg.in/src-d/go-git.v3/storage/proxy"
+	"gopkg.in/src-d/go-git.v3/storage/seekable"
 	"gopkg.in/src-d/go-git.v3/utils/fs"
 
 	"github.com/alcortesm/tgz"
@@ -19,7 +19,7 @@ var dirFixtures = [...]struct {
 }{
 	{
 		name: "binrels",
-		tgz:  "storage/proxy/internal/gitdir/fixtures/alcortesm-binary-relations.tgz",
+		tgz:  "storage/seekable/internal/gitdir/fixtures/alcortesm-binary-relations.tgz",
 	},
 }
 
@@ -77,7 +77,7 @@ func (s *SuiteRepository) TestNewRepositoryFromFS(c *C) {
 		c.Assert(err, ErrorMatches, `unable to find remote "origin"`)
 
 		c.Assert(repo.Storage, NotNil, com)
-		c.Assert(repo.Storage, FitsTypeOf, &proxy.ObjectStorage{}, com)
+		c.Assert(repo.Storage, FitsTypeOf, &seekable.ObjectStorage{}, com)
 	}
 }
