+++ storage/seekable/internal/gitdir/fixtures/empty-gitdir.tgz
