+++ storage/seekable/internal/gitdir/fixtures/no-packfile-no-idx.tgz
