+++ storage/seekable/internal/gitdir/fixtures/ref-deltas-no-idx.tgz
