+++ storage/seekable/internal/gitdir/fixtures/spinnaker-gc.tgz
