+++ storage/seekable/internal/gitdir/fixtures/spinnaker-no-idx.tgz
